requirements.txt
matplotlib==3.10.3
networkx==3.3
numpy==2.3.1
Pillow==11.2.1
scikit_learn==1.7.0
torch==2.6.0
torch_geometric==2.6.1
torchvision==0.21.0
